requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2
django-allauth==0.61.1
django-colorfield==0.11.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-widget-tweaks==1.5.0
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
urllib3==2.2.1

djangorestframework~=3.15.1